% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_simple_gmm.R
\name{demo_normal_performance}
\alias{demo_normal_performance}
\title{Shows KSD p value change with respect variation in noise}
\usage{
demo_normal_performance()
}
\description{
We generate a standard normal distribution,
and add varying gaussian noise to this dataset and see the change in pvalues.
}
