% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_simple_gmm.R
\name{demo_simple_gamma}
\alias{demo_simple_gamma}
\title{Tests 1-dimensional Gamma Distribution with customized parameters}
\usage{
demo_simple_gamma(
  trueshape = 10,
  truescale = 3,
  noisemu = 5,
  noisesd = 2,
  n = 100
)
}
\arguments{
\item{trueshape}{shape of true gamma distribution}

\item{truescale}{scale of true gamma distribution}

\item{noisemu}{mean of gaussian noise to add}

\item{noisesd}{standard deviation of gaussian noise to add}

\item{n}{number of samples to generate}
}
\description{
We generate a gamma distribution with given parameters,
and add gaussian noise to this dataset. We then compute the score
of each dataset for the original true distribution.
}
