% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_simple_gmm.R
\name{demo_simple_gaussian}
\alias{demo_simple_gaussian}
\title{Tests 1-dimensional Gaussian Distribution with customized parameters}
\usage{
demo_simple_gaussian(truemu = 5, truesd = 1, noisemu = 0, noisesd = 2, n = 100)
}
\arguments{
\item{truemu}{mean of true distribution}

\item{truesd}{standard deviation of true distribution}

\item{noisemu}{mean of gaussian noise to add}

\item{noisesd}{standard deviation of gaussian noise to add}

\item{n}{number of samples to generate}
}
\description{
We generate a gaussian distribution with given parameters,
and add noise to this dataset. We then compute the score
of each dataset for the original true distribution.
}
