% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{perturbgmm}
\alias{perturbgmm}
\title{Returns a perturbed model of given GMM}
\usage{
perturbgmm(model = NULL)
}
\arguments{
\item{model}{: The base Gaussian Mixture Model}
}
\value{
perturbedModel : Perturbed model with added noise to the supplied GMM
}
\description{
Returns a perturbed model of given GMM
}
\examples{
#Add noise to default 1-d gaussian mixture model
model <- gmm()
noisymodel <- perturbgmm(model)
}
